don't try to put size_t into tokenbuffer
would break compilation on macOS

+++ src/kernel/comsg/PROXY.cpp
@@ -17,10 +17,10 @@ namespace covise
 {
   IMPL_MESSAGE_WITH_SUB_CLASSES(PROXY, PROXY_TYPE)
   IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, CreateControllerProxy, int, vrbClientID)
-  IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, CreateSubProcessProxie, size_t, procID, sender_type, senderType, int, timeout)
+  IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, CreateSubProcessProxie, uint32_t, procID, sender_type, senderType, int, timeout)
   IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, ProxyCreated, int, port)
   IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, ProxyConnected, bool, success)
-  IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, CreateCrbProxy, size_t, toProcID, size_t, fromProcID, int, timeout)
+  IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, CreateCrbProxy, uint32_t, toProcID, uint32_t, fromProcID, int, timeout)
   IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, ConnectionTest, int, fromClientID, int, toClientID, int, port, int, timeout)
   IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, ConnectionState, int, fromClientID, int, toClientID, ConnectionCapability, capability)
   IMPL_SUB_MESSAGE_CLASS(PROXY, PROXY_TYPE, ConnectionCheck, int, fromClientID, int, toClientID)
